requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
dnspython==2.7.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==11.1.0
pip==24.3.1
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
Werkzeug==3.1.2
django-stubs
djangorestframework-stubs